+++ src/cloud-element-templates/util/bindingTypes.js
@@ -21,6 +21,7 @@ export const ZEEBE_FORM_DEFINITION = 'zeebe:formDefinition';
 export const ZEEBE_SCRIPT_TASK = 'zeebe:script';
 export const ZEEBE_ASSIGNMENT_DEFINITION = 'zeebe:assignmentDefinition';
 export const ZEEBE_PRIORITY_DEFINITION = 'zeebe:priorityDefinition';
+export const ZEEBE_JOB_PRIORITY_DEFINITION = 'zeebe:jobPriorityDefinition';
 export const ZEEBE_AD_HOC = 'zeebe:adHoc';
 export const ZEEBE_TASK_SCHEDULE = 'zeebe:taskSchedule';
 export const ZEEBE_EXECUTION_LISTENER = 'zeebe:executionListener';
@@ -41,6 +42,7 @@ export const EXTENSION_BINDING_TYPES = [
   ZEEBE_SCRIPT_TASK,
   ZEEBE_ASSIGNMENT_DEFINITION,
   ZEEBE_PRIORITY_DEFINITION,
+  ZEEBE_JOB_PRIORITY_DEFINITION,
   ZEEBE_AD_HOC,
   ZEEBE_TASK_SCHEDULE,
   ZEEBE_EXECUTION_LISTENER,